dnp3: rename files within dnp3 module
