+++ src/Makefile.am
@@ -18,13 +18,13 @@ alert-debuglog.c alert-debuglog.h \
 alert-fastlog.c alert-fastlog.h \
 alert-prelude.c alert-prelude.h \
 alert-syslog.c alert-syslog.h \
-app-layer/dnp3/app-layer-dnp3.c app-layer/dnp3/app-layer-dnp3.h \
-app-layer/dnp3/app-layer-dnp3-objects.c app-layer/dnp3/app-layer-dnp3-objects.h \
-app-layer/dnp3/detect-dnp3.c app-layer/dnp3/detect-dnp3.h \
-app-layer/dnp3/output-json-dnp3.c app-layer/dnp3/output-json-dnp3.h \
-app-layer/dnp3/output-json-dnp3-objects.c app-layer/dnp3/output-json-dnp3-objects.h \
-app-layer/dnp3/util-lua-dnp3.c app-layer/dnp3/util-lua-dnp3.h \
-app-layer/dnp3/util-lua-dnp3-objects.c app-layer/dnp3/util-lua-dnp3-objects.h \
+app-layer/dnp3/parser.c app-layer/dnp3/parser.h \
+app-layer/dnp3/parser-objects.c app-layer/dnp3/parser-objects.h \
+app-layer/dnp3/detect.c app-layer/dnp3/detect.h \
+app-layer/dnp3/log.c app-layer/dnp3/log.h \
+app-layer/dnp3/log-objects.c app-layer/dnp3/log-objects.h \
+app-layer/dnp3/lua.c app-layer/dnp3/lua.h \
+app-layer/dnp3/lua-objects.c app-layer/dnp3/lua-objects.h \
 app-layer.c app-layer.h \
 app-layer-dcerpc.c app-layer-dcerpc.h \
 app-layer-dcerpc-udp.c app-layer-dcerpc-udp.h \
