+++ src/app-layer/dnp3/detect.c
@@ -21,13 +21,13 @@
 
 #include "detect.h"
 #include "detect-parse.h"
-#include "detect-dnp3.h"
+#include "./parser.h"
 #include "detect-engine.h"
 #include "detect-engine-mpm.h"
 #include "detect-engine-prefilter.h"
 #include "detect-engine-content-inspection.h"
 
-#include "app-layer-dnp3.h"
+#include "./detect.h"
 
 static int g_dnp3_match_buffer_id = 0;
 static int g_dnp3_data_buffer_id = 0;
