+++ src/app-layer/dnp3/lua.c
@@ -17,8 +17,8 @@
 
 #include "suricata-common.h"
 
-#include "app-layer-dnp3.h"
-#include "app-layer-dnp3-objects.h"
+#include "./parser.h"
+#include "./parser-objects.h"
 
 #ifdef HAVE_LUA
 
@@ -28,8 +28,8 @@
 
 #include "util-lua.h"
 #include "util-lua-common.h"
-#include "util-lua-dnp3.h"
-#include "util-lua-dnp3-objects.h"
+#include "./lua.h"
+#include "./lua-objects.h"
 
 /**
  * \brief Helper macro to push key and integer value onto a table.
