+++ src/detect-lua-extensions.c
@@ -70,7 +70,7 @@
 #include "util-lua-tls.h"
 #include "util-lua-ssh.h"
 #include "util-lua-smtp.h"
-#include "app-layer/dnp3/util-lua-dnp3.h"
+#include "app-layer/dnp3/lua.h"
 #include "detect-lua-extensions.h"
 
 static const char luaext_key_ld[] = "suricata:luadata";
