+++ src/output-json-alert.c
@@ -45,7 +45,7 @@
 #include "detect-reference.h"
 #include "detect-metadata.h"
 #include "app-layer-parser.h"
-#include "app-layer/dnp3/app-layer-dnp3.h"
+#include "app-layer/dnp3/parser.h"
 #include "app-layer-htp.h"
 #include "app-layer-htp-xff.h"
 #include "app-layer-ftp.h"
@@ -56,7 +56,7 @@
 #include "output.h"
 #include "output-json.h"
 #include "output-json-alert.h"
-#include "app-layer/dnp3/output-json-dnp3.h"
+#include "app-layer/dnp3/log.h"
 #include "output-json-dns.h"
 #include "output-json-http.h"
 #include "output-json-tls.h"
