+++ src/suricata.c
@@ -126,7 +126,7 @@
 #include "app-layer-smtp.h"
 #include "app-layer-modbus.h"
 #include "app-layer-enip.h"
-#include "app-layer/dnp3/app-layer-dnp3.h"
+#include "app-layer/dnp3/parser.h"
 #include "app-layer-smb.h"
 #include "app-layer-dcerpc.h"
